+++ lib/ash_kotlin_multiplatform/rpc/verifiers/verify_action_types.ex
@@ -320,7 +320,7 @@ defmodule AshKotlinMultiplatform.Rpc.Verifiers.VerifyActionTypes do
     field_name_mappings = get_resource_field_mappings(embedded_module)
 
     attributes = Ash.Resource.Info.public_attributes(embedded_module)
-    calculations = Ash.Resource.Info.public_calculations(embedded_module)
+    calculations = AshKotlinMultiplatform.Resource.Info.public_field_calculations(embedded_module)
     aggregates = Ash.Resource.Info.public_aggregates(embedded_module)
     relationships = Ash.Resource.Info.public_relationships(embedded_module)
 
@@ -361,7 +361,7 @@ defmodule AshKotlinMultiplatform.Rpc.Verifiers.VerifyActionTypes do
     field_name_mappings = get_resource_field_mappings(target_resource)
 
     attributes = Ash.Resource.Info.public_attributes(target_resource)
-    calculations = Ash.Resource.Info.public_calculations(target_resource)
+    calculations = AshKotlinMultiplatform.Resource.Info.public_field_calculations(target_resource)
     aggregates = Ash.Resource.Info.public_aggregates(target_resource)
     relationships = Ash.Resource.Info.public_relationships(target_resource)